+++ pkg/registry/registry.go
@@ -359,7 +359,7 @@ func forwardRequest(client *http.Client, bufferPool *sync.Pool, req *http.Reques
 	if err != nil {
 		return err
 	}
-	copyHeader(forwardReq.Header, req.Header)
+	httpx.CopyHeader(forwardReq.Header, req.Header)
 	forwardResp, err := client.Do(forwardReq)
 	if err != nil {
 		return err
@@ -372,7 +372,7 @@ func forwardRequest(client *http.Client, bufferPool *sync.Pool, req *http.Reques
 
 	// TODO (phillebaba): Is it possible to retry if copy fails half way through?
 	// Copy forward response to response writer.
-	copyHeader(rw.Header(), forwardResp.Header)
+	httpx.CopyHeader(rw.Header(), forwardResp.Header)
 	rw.WriteHeader(http.StatusOK)
 	//nolint: errcheck // Ignore
 	buf := bufferPool.Get().(*[]byte)
@@ -382,12 +382,4 @@ func forwardRequest(client *http.Client, bufferPool *sync.Pool, req *http.Reques
 		return err
 	}
 	return nil
-}
-
-func copyHeader(dst, src http.Header) {
-	for k, vv := range src {
-		for _, v := range vv {
-			dst.Add(k, v)
-		}
-	}
 }