+++ CHANGELOG.md
@@ -16,6 +16,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 - [#906](https://github.com/spegel-org/spegel/pull/906) Replace HTTP header strings with httpx constants.
+- [#916](https://github.com/spegel-org/spegel/pull/916) Refactor OCI client options and add header configuration.
 
 ### Deprecated